# Parallelize time-consuming parts when there are many local branches