I have more than 30 local branches working with.
This is partly because we are using Graphite.dev to make stacked, fine-grained, small-sized PRs, increasing the overall number of PRs.

After inserting simple `Println()` calls around the code, I found that `applyCommits()` and querying the PRs by commit hashes from the GitHub were the two major bottlenecks with a large number of local branches.

This PR simply parallelizes them using goroutines and wait groups.

### A simple benchmark:
**Before:**
```
gh poi  2.03s user 0.72s system 16% cpu 16.858 total
```
**After:**
```
gh poi  1.53s user 0.65s system 42% cpu 5.136 total
```

You could observe that the most of time were wasted by waiting for I/O.