@seachicken 2025-01-07
It's good that it will be faster. The reason I put off parallelization is that the GitHub API has a rate limit, and if there are too many requests, an error may occur. I'll check the specifications around that.