@achimnol 2025-01-08
> It's good that it will be faster. The reason I put off parallelization is that the GitHub API has a rate limit, and if there are too many requests, an error may occur. I'll check the specifications around that.

Yes, I had the same concern, too. Nonetheless, as the GitHub API queries for getting corresponding PRs for a set of commit hashes are [already chunked by 6 hashes due to the variable length limit (256 characters) in the GraphQL schema](https://github.com/seachicken/gh-poi/blob/67aa4cc1c62448d57a4c8a9197c446827dd03776/shared/querygen.go#L46). So, even with ~30 local branches, the number of actual GitHub API requests would be less than 10 for a single execution of `gh poi`.

I think practically this shouldn't be the issue as long as the user is logged in, unless the user repeats `gh poi` quickly in a short period of time.